refactor: improve webapp component with useAppTheme hook
